sys/net/aodvv2: move rfc5444 compat code to rfc5444
