+++ sys/net/aodvv2/aodvv2_writer.c
@@ -24,7 +24,6 @@
 #include "aodvv2_writer.h"
 #include "net/aodvv2/metric.h"
 
-#include "rfc5444_compat.h"
 
 #define ENABLE_DEBUG (0)
 #include "debug.h"
